Input script (first upload)
This is the first functional input script. Right now it is only being used to study the equilibration process when the 'hot' and 'cold' layers are set to 10K. The corresponding .xyz data dump shows the non-stationary layers moving quite vigorously, which makes me think something isn't quite right with this one.