Faithful input file
This input file has been made in an attempt to remain faithful to the setup presented in the paper. That is, I first try to equilibrate the system at 10K before applying NVT thermostats in order to fix the 'cold' group at 10K and raise the 'hot' group to 80K. However, the right-hand side of the argon layer (i.e. interfacing with the 'cold' group) seems hotter than it should be. Probably a sign of equilibration not having been reached. 
Also, using unfixes in order to remove NVE on hot and cold groups results in segfaults (the unfixing was done with an input file the more closely resembles the one that I will upload in the next commit).