Input file with 'forced 10K for Argon and cold'
In this input file, I skip using an equilibration run followed by an NVT run with hot at 80K. Instead, I sample hot with velocities corresponding to 80K and 10K for the rest straight away and apply NVT thermostats on hot and cold and NVE on the argon layer. This provides graph shapes (see bottom of .ods made on 04.08.17) that look vaguely like the paper's, but the scale, i.e. values of the temperature, particularly for the layers near 'hot', are incorrect (they're closer to 10K rather than 80K), even after a couple of ps of runs.

TO DO on Monday:
- Make sure equilibration involves putting all moving atoms at 10K (today I had mistakenly thought only hot and cold were meant to be at 10K during equilibration)
- Make argon thickness 20nm, will have to increase number of atoms included in a group as a result
- Produce graphs with real-unit scales