Input with compute temp/region
This input file uses the compute temp/region command in order to avoid creating separate groups for each Argon layer in the film. Instead, regions are constructed for a single Argon film group. This made it possible to create 60 temperature computes that should span the whole film length - which is believed to be approximately 20 nm.

Note, this file ONLY equilibrates the Argon layer at the moment. I have not yet tried raising the hot region's temperature and applying new thermostats. That is, it currently:
- Sets velocities of non-fixed atoms using Gaussian distribution at 10K
- NVT thermostats the hot and cold regions at 10K
- Runs for 2000 timesteps and thermodynamic report gives temperature of each argon layer region every 500 timesteps

To investigate: after 500 timesteps, the temperature of the first argon layer is higher than the rest, why?