Updated temp/region - faithful attempt
This input file initialises Argon atoms at 10K using the velocity command, performs NVE integration, then unfixes the NVEs and follows up with NVT integration.

Note: the temperature of cold is specified as 0.04 units - this is ad-hoc. It is only done because equilibration always seem to make the temperature of all the argon atoms drop from 0.08 to 0.04. The temperature profile of the argon 18 ps after the start of the NVT phase resembles the paper's graph - but the values are completely wrong.

To fix:
- Temperature should stay at 10K during equilibration
- Temperature should not rise so dramatically during NVT phase - investigate why. Is the double integration of certain atoms causing this?