Fixed tdamp,ad-hoc Tstart for NVT
This input file now utilises 4 atom types (identical in terms of physical properties) assigned to different groups, so that the group command no longer uses the region keyword. This removes the double integration warning found earlier.

Currently, the process is still moderately faithful:
(1) Atoms initialised with velocities corresponding to 10K Gaussian distribution
(2) System is NVE integrated for 6 ps (since nanosecond simulation takes a very long time)
(3) NVEs for hot and cold are removed and replaced with NVT fixes. Given that the temperature after equilibration pretty much always reaches 5K, I have set the starting temperatures for both NVTs at 5K (0.04 in LJ units), and the final temperature for hot is set to 80K (0.6608) whilst the final temperature for cold is the same as initial. The system is run for 100 timesteps (an attempt to make hot rise in temperature rapidly).
(4) Hot's NVT is modified so that both starting and final temperatures are now at 0.6608 a