New density; self/spring
I have changed the reduced density, given that the atoms in the centres of the cubic faces are positioned closer to corner atoms than the lattice spacing. Hence, setting the lattice spacing to the equilibrium distance would place the face atoms within the positive part of the LJ potential, leading to exponentially large, positive total and potential energies.

A self/spring command was added on the advice of Karthik. This is essentially to ensure that the Argon behaves like a solid, because it is suggested that the LJ potential is too weak otherwise (LJ is usually used for gases). We think that the phonon dynamics should not be affected too much, since phonon transmission should depend primarily on interatomic interactions. 

The spring constant is chosen such that the potential and total energy does not fluctuate very much. In Karthik's script, we found this to be 1000 N/m, and so this is the value I have chosen (converted into LJ units).

To investigate:
- Use a Langevin the