Increased atoms (3040 atoms)
Increased breadth and width of box (and also changed overall simulation box parameters so that it should always be large enough to contain the atoms inside it) so that no. of atoms increased from ~700 to 3040. Note, paper makes use of a maximum of approx 5000 atoms.

The resultant temperature profile seems to resemble the paper's one closely, barring the presence of a mechanical wave. Will probably need to make the temperature raise phase shorter and add a variable that subtracts the mean velocity of atoms in the layers in order to see it.